requirements.txt
acora==2.2
adblockparser==0.7
AdvancedHTTPServer==2.2.0
aiocmd==0.1.2
aioconsole==0.3.1
aiodns==2.0.0
aiohttp==3.6.2
aiomultiprocess==0.8.0
aioredis==1.3.1
aiosqlite==0.16.1
aiowinreg==0.0.3
ajpy==0.0.4
alembic==1.4.3.dev0
aniso8601==8.1.1
apispec==3.3.1
apispec-webframeworks==0.5.2
appdirs==1.4.4
APScheduler==0.0.0
argon2-cffi==18.3.0
arrow==0.17.0
artifacts==20201106
asciinema==2.0.2
asciitree==0.3.3
asgiref==3.3.0
asn1crypto==1.4.0
async-timeout==3.0.1
asysocks==0.0.2
attrs==20.3.0
autobahn==17.10.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
basemap==1.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
bibtexparser==1.1.0
binwalk==2.2.0
biom-format==2.1.10
bleach==3.2.1
blinker==1.4
boltons==19.1.0
bottle==0.12.19
Bottleneck==1.2.1
Brotli==1.0.9
CacheControl==0.12.6
cairocffi==1.2.0
capstone==4.0.2
catfish==4.16.0
cbor==1.0.0
certifi==2020.6.20
cffi==1.14.4
chardet==4.0.0
CherryPy==8.9.1
chrome-gnome-shell==0.0.0
click==7.1.2
click-plugins==1.1.1
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
construct==2.10.58
crackmapexec==5.1.4.dev0
cryptography==3.3.2
cupshelpers==1.0
cutadapt==3.2
cwiid==0.6.91
cycler==0.10.0
Cython==0.29.21
dblatex==0.3.12
dbus-python==1.2.16
debtags==2.1
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.0.7
dicttoxml==1.7.4
distro==1.5.0
distro-info==1.0
Django==2.2.19
dnaio==0.5.0
dnslib==0.9.14
dnspython==2.0.0
docopt==0.6.2
ecdsa==0.16.1
elevate==0.1.3
email-validator==1.1.1
entrypoints==0.3
et-xmlfile==1.0.1
evdev==1.4.0
ExifRead==2.3.2
expiringdict==1.1.4
faraday-client==1.0.0
faraday-plugins==1.2
faradaysec==3.14.1
feedparser==5.2.1
fierce==1.4.0
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.5
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-KVSession-fork==0.6.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security-Too==4.0.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.18.2
gattlib==0.20201113
GDAL==3.2.1
geoip2==2.9.0
geojson==2.5.0
gitdb==4.0.5
GitPython==3.1.12
gpg===1.14.0-unknown
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
h11==0.12.0
h2==4.0.0
h5py==2.10.0
h5py.-debian-h5py-serial==2.10.0
hashID==3.1.4
hdmedians==0.14.1
hiredis==1.0.1
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
humanize==0.0.0
hupper==1.10.2
hyperframe==6.0.0
hyperlink==19.0.0
icalendar==4.0.3
idna==2.10
ifaddr==0.1.7
ijson==3.1.3
impacket==0.9.22
importlib-metadata==1.6.0
incremental==17.5.0
iniconfig==0.0.0
intervaltree==3.0.2
invoke==1.4.1
ipwhois==0.15.1
IPy==1.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kaitaistruct==0.9
kamvas-driver==0.1.2
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kiwisolver==1.3.1
ldap3==2.8.1
ldapdomaindump==0.9.3
lightdm-gtk-greeter-settings==1.2.2
lockfile==0.12.2
lsassy==2.1.2
lxml==4.6.2
lz4==3.1.3+dfsg
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.19.0
mate-menu==20.4.1
matplotlib==3.3.4
maxminddb==2.0.3
mechanize==0.4.5
minidump==0.0.12
minikerberos==0.2.0
mistune==0.8.4
mitmproxy==6.0.2
more-itertools==4.2.0
msgpack==1.0.0
msldap==0.2.10
multidict==5.1.0
mysqlclient==1.4.4
nassl==4.0.0
natsort==7.1.0
nbconvert==5.6.1
nbformat==5.1.2
neo4j==1.7.0.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.7.19
netifaces==0.10.9
networkx==2.5
nose==1.3.7
notebook==6.2.0
nplusone==1.0.0
ntlm-auth==1.4.0
numexpr==2.7.2
numpy==1.19.5
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parsedatetime==2.6
parso==0.8.1
passlib==1.7.2
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
patator==0.9
patsy==0.5.1
pcapy==0.11.4
pefile==2019.4.18
pexpect==4.8.0
pgspecial==1.11.10
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==8.1.0
plaster==1.0
plaster-pastedeploy==0.5
plotly==4.14.3
pluggy==0.13.0
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
prometheus-client==0.9.0
promise==2.2
prompt-toolkit==3.0.14
protobuf==3.12.4
psutil==5.8.0
psycopg2==2.8.6
publicsuffix2==2.20191221
py==1.10.0
py-ubjson==0.16.1
pyaff4==0.26.post6
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyBluez==0.23
pycairo==1.16.2
pycares==3.1.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pydot==1.4.1
pyee==7.0.2
pyelftools==0.27
pyExploitDb==0.2.0
pyfiglet==0.8.post0
Pygments==2.7.1
PyGObject==3.38.0
pygraphviz==1.6
PyHamcrest==1.9.0
pyinotify==0.9.6
pylibiio==0.21
pylnk3==0.2.1
pymssql==2.1.4
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyotp==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.0
pypng==0.0.20
pyppeteer==0.2.5
pyproj==3.0.0
pypsrp==0.5.0
pypykatz==0.3.7
PyQRCode==1.2.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
pyramid==1.10.4
pyrsistent==0.15.5
pyserial==3.5b0
pyShodan==0.2.3
pyshp==2.1.3
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pyspnego==0.1.5
pytest==6.0.2
python-apt==2.1.7
python-dateutil==2.8.1
python-debian==0.1.39
python-docx==0.8.10
python-editor==1.0.3
python-magic==0.4.20
python-miio==0.5.0.1
python-pam==1.8.4
python-pptx==0.6.18
python-snappy==0.5.3
python-xlib==0.29
PyTrie==0.2
pytsk3==20200117
pytz==2021.1
pyudev==0.22.0
pyusb==1.1.1
pywerview==0.2.0
pyxdg==0.27
PyYAML==5.3.1
pyzmq==20.0.0
q2-cutadapt==2020.11.1
q2-demux==2020.11.1
q2-feature-classifier==2020.11.1
q2-feature-table==2020.11.1
q2-metadata==2020.11.1
q2-quality-control==2020.11.1
q2-quality-filter==2020.11.1
q2-taxa==2020.11.1
q2-types==2020.11.1
q2cli==2020.11.1
q2templates==2020.11.1
qiime2==2020.11.1
qrcode==6.1
Quamash==0.6.1
rdflib==5.0.0
redis==3.5.3
rekall-core==1.7.2rc1
rekall-lib==1.7.1
repoze.lru==0.7
requests==2.25.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
retrying==1.3.3
Routes==2.5.1
rq==1.7.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
rule-engine==1.1.0
Rx==3.1.1
scapy==2.4.4
scikit-bio==0.5.6
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.11.1
secure==0.2.1
selenium==4.0.0a1
Send2Trash==1.6.0b1
service-identity==18.1.0
setproctitle==1.2.1
shodan==1.25.0
simplejson==3.17.2
simplekv==0.13.0
six==1.15.0
smmap==3.0.4
smoke-zephyr==2.0.1
sortedcontainers==2.1.0
soupsieve==2.2
SPARQLWrapper==1.8.5
speaklater==1.4
SQLAlchemy==1.3.22
sqlalchemy-schemadisplay==1.3
sqlparse==0.4.1
sslyze==4.0.4
stem==1.8.0
syslog-rfc5424-formatter==1.2.2
tables==3.6.1
tabulate==0.8.7
Tempita==0.5.2
termcolor==1.1.0
terminado==0.9.2
terminaltables==3.1.0
testpath==0.4.4
texttable==1.6.3
theHarvester==3.2.3
tld==0.11.11
tls-parser==1.2.2
toml==0.10.1
torbrowser-launcher==0.3.3
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
translationstring==1.4
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
u-msgpack-python==2.3.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.2
urwid==2.1.2
uvloop==0.14.0
venusian==3.0.0
wafw00f==2.1.0
wapiti3==3.0.3
wcwidth==0.1.9
webargs==7.0.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wfuzz==3.1.0
whois==0.8
widgetsnbextension==2.0.0
wifite==2.5.8
winacl==0.0.2
wsaccel==0.6.2
wsproto==0.15.0
WTForms==2.2.1
xcffib==0.8.1
xdg==5
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmltodict==0.12.0
xopen==1.1.0
yara-python==4.0.4
yarl==1.6.3
yaswfp==0.9.3
zeroconf==0.26.1
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==5.2.0
